+++ scripts/schema/db/init_dbs/postgresql/init_schema.sql
@@ -1117,7 +1117,7 @@ $$
             
             CREATE TABLE public.tags
             (
-                tag_id               bigint       NOT NULL PRIMARY KEY,
+                tag_id               serial       NOT NULL PRIMARY KEY,
                 name                 text         NOT NULL,
                 project_id           integer      NOT NULL REFERENCES public.projects (project_id) ON DELETE CASCADE,
                 selector             text         NOT NULL,
